make reasource controller usable

+++ src/controllers/resource.controller.ts
@@ -1,66 +1,82 @@
-import {Filter, Where, repository} from '@loopback/repository';
+import { Filter, Where, repository } from '@loopback/repository';
 import {
   post,
   param,
   get,
   patch,
   del,
-  requestBody
+  requestBody,
+  HttpErrors
 } from '@loopback/rest';
-import {Resource} from '../models';
-import {ResourceRepository} from '../repositories';
+import { Resource, User } from '../models';
+import { ResourceRepository } from '../repositories';
+import {
+  AuthenticationBindings,
+  authenticate,
+} from '@loopback/authentication';
+import { inject, Getter } from '@loopback/context'
 
 export class ResourceController {
   constructor(
     @repository(ResourceRepository)
-    public resourceRepository : ResourceRepository,
-  ) {}
+    public resourceRepository: ResourceRepository,
+    @inject.getter(AuthenticationBindings.CURRENT_USER)
+    private getCurrentUser: Getter<User>,
+  ) { }
 
-  @post('/resources')
-  async create(@requestBody() obj: Resource)
-    : Promise<Resource> {
-    return await this.resourceRepository.create(obj);
-  }
-
+  @authenticate('JwtStrategy')
   @get('/resources/count')
   async count(@param.query.string('where') where?: Where): Promise<number> {
+    const currentUser = await this.getCurrentUser();
+    if (!currentUser.isAdmin) {
+      throw new HttpErrors.Forbidden('Access denied')
+    }
+
     return await this.resourceRepository.count(where);
   }
 
+  @authenticate('JwtStrategy')
   @get('/resources')
   async find(@param.query.string('filter') filter?: Filter)
     : Promise<Resource[]> {
+    const currentUser = await this.getCurrentUser();
+    if (!currentUser.isAdmin) {
+      throw new HttpErrors.Forbidden('Access denied')
+    }
+
     return await this.resourceRepository.find(filter);
   }
 
-  @patch('/resources')
-  async updateAll(
-    @requestBody() obj: Resource,
-    @param.query.string('where') where?: Where,
-  ): Promise<number> {
-    return await this.resourceRepository.updateAll(obj, where);
-  }
-
+  @authenticate('JwtStrategy')
   @del('/resources')
   async deleteAll(@param.query.string('where') where?: Where): Promise<number> {
+    const currentUser = await this.getCurrentUser();
+    if (!currentUser.isAdmin) {
+      throw new HttpErrors.Forbidden('Access denied')
+    }
+
     return await this.resourceRepository.deleteAll(where);
   }
 
+  @authenticate('JwtStrategy')
   @get('/resources/{id}')
   async findById(@param.path.number('id') id: number): Promise<Resource> {
+    const currentUser = await this.getCurrentUser();
+    if (!currentUser.isAdmin) {
+      throw new HttpErrors.Forbidden('Access denied')
+    }
+
     return await this.resourceRepository.findById(id);
   }
 
-  @patch('/resources/{id}')
-  async updateById(
-    @param.path.number('id') id: number,
-    @requestBody() obj: Resource
-  ): Promise<boolean> {
-    return await this.resourceRepository.updateById(id, obj);
-  }
-
+  @authenticate('JwtStrategy')
   @del('/resources/{id}')
   async deleteById(@param.path.number('id') id: number): Promise<boolean> {
+    const currentUser = await this.getCurrentUser();
+    if (!currentUser.isAdmin) {
+      throw new HttpErrors.Forbidden('Access denied')
+    }
+
     return await this.resourceRepository.deleteById(id);
   }
 }
